move summary query tests
